// frontend/src/backend/cartly/schema.d.ts
export interface components {
    schemas: {
        /** CategoryCreate */
        CategoryCreate: {
            /** Name */
            name: string
            /** Color */
            color?: string | null
            /** Icon */
            icon?: string | null
        }
        /** CategoryResponse */
        CategoryResponse: {
            /**
             * Id
             * Format: uuid
             */
            id: string
            /** Name */
            name: string
            /** Color */
            color?: string | null
            /** Icon */
            icon?: string | null
        }
        /** CategoryUpdate */
        CategoryUpdate: {
            /** Name */
            name: string
            /** Color */
            color?: string | null
            /** Icon */
            icon?: string | null
        }
        /** HTTPValidationError */
        HTTPValidationError: {
            /** Detail */
            detail?: components['schemas']['ValidationError'][]
        }
        /** ItemCreate */
        ItemCreate: {
            /** Name */
            name: string
            /**
             * Quantity
             * @default 1
             */
            quantity: number
            /** Category Id */
            category_id?: string | null
        }
        /** ItemResponse */
        ItemResponse: {
            /**
             * Id
             * Format: uuid
             */
            id: string
            /** Name */
            name: string
            /** Quantity */
            quantity: number
            /** Is Checked */
            is_checked: boolean
            /**
             * Shopping List Id
             * Format: uuid
             */
            shopping_list_id: string
            /** Category Id */
            category_id?: string | null
            /**
             * Created At
             * Format: date-time
             */
            created_at: string
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string
        }
        /** ItemUpdate */
        ItemUpdate: {
            /** Name */
            name: string
            /**
             * Quantity
             * @default 1
             */
            quantity: number
            /**
             * Is Checked
             * @default false
             */
            is_checked: boolean
            /** Category Id */
            category_id?: string | null
        }
        /** ShoppingListCreate */
        ShoppingListCreate: {
            /** Name */
            name: string
            /** Description */
            description?: string | null
        }
        /** ShoppingListResponse */
        ShoppingListResponse: {
            /**
             * Id
             * Format: uuid
             */
            id: string
            /** Name */
            name: string
            /** Description */
            description?: string | null
            /**
             * Created At
             * Format: date-time
             */
            created_at: string
            /**
             * Updated At
             * Format: date-time
             */
            updated_at: string
        }
        /** ShoppingListUpdate */
        ShoppingListUpdate: {
            /** Name */
            name: string
            /** Description */
            description?: string | null
        }
        /** ValidationError */
        ValidationError: {
            /** Location */
            loc: (string | number)[]
            /** Message */
            msg: string
            /** Error Type */
            type: string
            /** Input */
            input?: unknown
            /** Context */
            ctx?: Record<string, never>
        }
    }
    responses: never
    parameters: never
    requestBodies: never
    headers: never
    pathItems: never
}